New system to automatically update version strings
The index.php file specifies a version query string for each
resource which is likely to change. For example:

    <script src="/js/sim.js?v0.8.2c"></script>

Each time the file is changed, we have incremented the version
number so that browsers and/or the CDN know to download the
file from the web site, rather than using a cached version.
This works because changing the version query string changes
the file name of the resource.

This commit introduces a new program (update-version-query-strings)
which parses index.php and sets each version query string to be
the MD5 hash of the corresponding resource. On the live web
server, this program is set to run for the following git hooks:

    post-commit, post-checkout, post-merge, post-rewrite

This new system removes the need to manually update version query
strings. Instead, the version query strings will be automatically
updated when we update the live web site from now on.