Implement timestamps; save prefs in localStorage
This commit implements timestamp functionality in the lobby chat,
in private messages, and in battle chats. Timestamps are controlled
by a /timestamps command, which has three legal invocations:

    /timestamps off
        turns off timestamps

    /timestamps minutes
        show timestamps of the form [hh:mm]

    /timestamps seconds
        show timestamps of the form [hh:mm:ss]

For now, the default setting is /timestamps minutes. If this proves
unpopular, it can be changed. The timestamp preference is stored in
localStorage and does not have to be set every time.