Don't load the client if teams fail to load
If teams fail to load due to some bug, the user may lose their teams
permanently if they save their teams. In avoid to avoid a catastrophe,
we now decline to load the client if teams fail to load.

This should never happen unless a bug is introduced that breaks team
loading.