Sharing of resources between psim.us and pokemonshowdown.com
This commit allows the new psim.us domain to use the cookies and
`localStorage` of play.pokemonshowdown.com. Teams and prefs are
initially loaded from the play.pokemonshowdown.com `localStorage` and
changes to teams and prefs are written back to the
play.pokemonshowdown.com `localStorage`.

This commit also implements a few other things for the new psim.us
domain (implemented by Zarel):

- URIs of the form serverid.psim.us load the server `serverid`
- the naked domain now loads the lobby