Split replayer.css replayer.css was sort of a badly defined file: it started out as being all the CSS used by the replay viewer, except for the CSS for battles. It then gained utilichart CSS, despite being mostly unrelated.
Now, main layout is in main.css in the main domain, and replay
layout is in replay.css, and utilichart CSS is in utilchart.css.

replayer.css remains a legacy file for the old replay player.
Once we transition to the new replay player that uses replay.css,
I can get rid of replayer.css completely.