Connection: Fix alt port support when connecting over HTTPS
If the client fails to establish an HTTPS connection to the main server,
the page needs to be re-loaded over HTTP before attempting to connect
on an alternative port; otherwise, there will be mixed content
security errors.

I had actually implemented this feature in the old (pre-v0.90) client,
but I forgot to copy it into the new client when I implemented the
connection code.