Changed the behavior of /challenge /challenge now opens the challenge window instead of the user window (like /user does)
/challenge also accepts multiple parameters now: user, format, team
user is the name(id) of the user you want to challenge
format is the format you want. It gets changed to an id, so you can enter it in any way you want. For example: ou, ou no stealth rock, etc.
team is the name of the team you wish to use. It also gets changed to an id and stops at the first match (problem: multiple teams with the same name?)
Both the format and team parameters are optional