Suppress crash on move request with all actives fainted
There's a weird bug in which it's possible to open a choose-move
screen when the client believes that all your active Pokemon are
fainted (this is not true). It seems to involve having a team
with multiple Pokemon with the same name, multiple Pokemon
with the same species, and/or Illusion (in case it was unclear,
this is a Hackmons-specific bug).

I don't have time to look at it in depth, but this commit
suppresses it enough to make PS playable instead of crashing
in that situation.