Implement old gen sprites for past gens
Implement old gens to use the old sprites as well as an option to disable it that may interact with enable BW sprites.
If old gens are not disallowed, those will be used for past gens (1-5).
If BW sprites are allowed, those will be used for XY and BW.
If old gens are disallowed and BW allowed, all gens will use BW sprites.