Auto-link URLs without closing parenthesis at end
Currently Pokémon Showdown doesn't auto-link URLs properly in the
following form.

    (Check out http://play.pokemonshowdown.com/)

In this case, it's obvious that `)` not part of an URL, but the current
auto-linker believes that it is. Making `)` illegal character at end of
URL doesn't solve the problem, as sometimes URLs actually do contain
parenthesis. The new URL regex supports one level of parenthesis
(regexes don't support recursion, but if you have more than one layer
of parenthesis in an URL, you are doing something wrong). This is a
case with URLs on Wikipedia.

    https://en.wikipedia.org/wiki/Pokémon_(disambiguation)