Fix links to PS rooms
Clicking links to PS rooms can be pretty unreliable in terms of whether
and where they open. This commit uses a global click handler, to make
sure that any and every link to PS rooms, including links in the News,
will open in the current PS window.

You can always right-click and Open in New Tab, of course.