Open Rooms room when it fits
By default, the Join Room window is only auto-opened on startup if
it fits. This changes it so that if it doesn't fit on startup but
the window is later resized to fit, the Join Room window will
auto-open if it hasn't been opened yet.

In addition to behaving more expectedly, this deals with an issue
in NW.js in Windows where maximizing doesn't happen before PS
loads.