Desktop: Dock badge notifications
Notification count is now displayed on the Dock icon in the desktop
client (in Windows, the taskbar needs to be in Large Icons mode).

Since notifications aren't persistent, this is useful as a long-term
reminder that you have something in PS pending.

This uses a newer NW.js feature, so it requires updating to the
latest version of the PS desktop client (0.3.0).