Allow negative attack nature for U-turn and Volt Switch
U-turn and Volt Switch are unusual in an aspect that they are dual
purpose moves. Sometimes they are used as damaging moves (main
attacking move even, like with Mega Beedrill or Genesect). In other
cases, they are used to preserve momentum, and the damage from them
doesn't exactly matter (see Cobalion, Xatu, Mega Pidgeot or
Hydreigon).

This change allows giving a negative attacking nature to an user of
U-turn or Volt Switch if it appears to be used in a momentum-sort of
way, as opposed to an usage as a damaging move.