Add subtle notifications for room activity
Chat room titles now turn orange and their icon changes when someone
talks in them. This is subtler than the highlight notification, and
should make it easier to keep up with rooms.

This change involves a refactor of the notification system, which
also happens to improve room-switching performance by only running
updateTopbar once instead of 3-5 times like it would before.