Improve battle protocol support
The battle protocol now supports leading and trailing spaces. This
drops support for the previous battle protocol which had spaces
between pipes, but I'm pretty sure this is safe since the replay
saving feature was not implemented until after that change.

This feature is mainly useful for ensuring that the first character
of a username is the user's symbol.