Move index.template.html character set declaration to HTTP
According to [WHATWG specification][1], there are a few of requirements
that a character set declaration must meet. One of these is a following
one.

> The element containing the character encoding declaration must be
> serialised completely within the first 1024 bytes of the document.

Beginning of a document contains a **huge** comment (pretty much 2
kilobytes). Either it would have to be moved to be after character
set declaration, character set declaration would have to be moved
to HTTP layer, or a comment would have to be remove entirely. I decided
on a second option, because I couldn't think of a way to move this
comment without making it uglier.

[1]: https://html.spec.whatwg.org/multipage/semantics.html#charset