Automatic autojoin
Basically, when opening PS, reopen the last chatrooms that were open
when PS was closed.

Limit 8 rooms. Staff and Upper Staff don't count towards the limit
because they're autojoined server-side. Lobby is not allowed to be
autojoined for server load reasons, and because it's autojoined
server-side on many other servers.

Currently only works for PS Main. An update that works for other
servers is coming soon.