Roomlist refactor
This adds the 'Join other room' button on first load, allowing it to
be focused before the room list is retrieved. The button is also
focused whenever the room is focused (by clicking inside the room
or by switching to the room's tab), allowing people to switch out
of the room by pressing Left, or opening the join menu by pressing
Enter.