Improve spoiler tag support
Putting parentheses like this: (spoiler: test) now allows you to
designate a part of a line to be spoilered, instead of spoilering
everything to end-of-line. Yes, it's also possible to have
multiple spoilers on a line.

As a side effect, this also fixes the issue where spoilers messed
up modlogs by restricting them to one line.