Don't close when forfeiting from Close button
The Forfeit button in the menu that appears when you try to Close an
active battle is actually a 'Forfeit and close' button. It's now
just a Forfeit button, in case players want to stay in the room
for whatever reason. Actually leaving the room just involves clicking
the Close button again.