Drag-rearrange-able teams
This is pretty experimental, but all the bugs I found are worked
out.

While researching how to best implement this, I came across the
famous PPK rant on why HTML5 drag-and-drop sucks:
http://quirksmode.org/blog/archives/2009/09/the_html5_drag.html

A lot of his complaints aren't really a big deal to me, but boy
was he right about browser bugs. Apparently mobile will never
support HTML5 drag-and-drop, either, and there are a lot of
other limitations. Oh well, this is all worth it for an
upcoming planned features: dragging and dropping team files to
and from your desktop.