Nicer-looking tab overflow menu
.button styling has been consolidated in CSS, and the tab overflow
menu has been adjusted to use button styling, which is the "more
improvements in the future" mentioned by 69524e935b.

I now declare the tab overflow menu officially out of beta. Now
for the rest of PS!

The Ladder now also uses the new button styling, but it's not out
of beta yet, I just wanted to see how it looked.