Fix crashes in base power display
Previously, any time there was only one opposing Pokemon remaining in
Doubles/Triples and it wasn't the leftmost one (viewed from the back),
the tooltip for any of the moves requiring a target would crash when
moused over. This basically caused an automatic forfeit for those
playing on the desktop app.