Better support for 'working offline'
PS's UI is now updated to make it clearer whether or not you're
connected to a server, and which features are and aren't
available while disconnected.

The Reconnect? popup's "close" button has been renamed "Work offline",
to make it clearer that it's for people who want to continue using
the teambuilder or whatever while disconnected.

The button is a slight misnomer since some features, like the
Ladder, don't work without an internet connection.