Better fuzzy matching in teambuilder search
If no exact match is found, the matcher is now guaranteed to display
an approximate match with the most shared starting letters with the
query.

In other words, 'assv' will now show 'Assault Vest' as the first
option.

In addition, we still search for other matches when there's an
alias match. So 'zard'->'charizard' can match 'Charizard-Mega-Y',
but 'ferro'->'ferrothorn' can still match 'Ferroseed'.

This code is getting kind of messy, but oh well.