Change close-roomtab button to button
There's a reason why close-roomtab buttons are <a> tags right now, and
I think it has to do with compatibility with some obscure browser like
IE9 or something.

Still, though, the CSS is much nicer this way, especially for the
upcoming dark mode feature, so I'll just commit this and figure out a
different workaround for whatever problems come up.