Fix colors for Dark Mode
After a bunch of iterations, this color scheme mostly only has blue
and yellow users complaining.

Blue names are now lighter, and yellow names are now darker, so that
they're more readable in dark mode and regular mode respectively.

This is done using a fancy algorithm to figure out the luminosity
of the generated color, and adjusting accordingly.