Use favicon-notify for unread notifications
I actually made favicon-notify around 4 years ago, even predating when
we moved to Git for the client repo.

Unfortunately, at the time, a Chrome bug made it not work, so I gave up
on it back then and decided web notifications would be enough.

I've missed enough notifications to know that having both would be
better, and all the relevant browser bugs have been fixed so this can
be implemented the easy way.