Make |unlink|hide| always roomhide
|unlink|hide| used to hide all text from the user everywhere, not
just in the room. While a good idea in spirit, in practice it
was inconsistent in terms of what it hid, because not all users
are in the same room, etc etc. It also made for very confusing
behavior when text would disappear without a message, when a user
was locked in a different room.

This also fixes bugs caused by this behavior.