Don't notify for own messages
For the most part, this isn't really a big deal: If you send a message,
you usually have the window open, so you can see that you've sent a
message.

This is mainly for things like if you send a message and switch rooms
before it appears in the roomlog, but it's really mainly for things
like /hotpatch where you don't need to be notified about all the
rooms that notify you.