Autojoin regardless of window width
This fixes autojoin not working because of the window width being too narrow.
This was from before chatrooms besides Lobby were implemente, but nowadays it
depends on the window layout of a desktop giving the browser enough width,
isn't consistent with autojoining via routing or server-side autojoin, and for
mobile users makes the function dependent on the screen resolution and whether
or not the phone's in profile or landscape mode.

Fixes #742