Battles: Fix bug in team preview
Previously, a race condition would make the team preview selector appear
before we knew how many pokemon needed to be selected (e.g. 4 in VGC).

This new UI patches over that problem, although the real fix (mentioned
in the TODO comment) is to send that information in the request object.