Make front sprites 1.5x size in gen 1-4
PS makes back sprites larger than front sprites, which is standard for
Gen 5 (although we do it slightly differently - Gen 5 makes backsprites
2x size; we make backsprites 1.5x size) and onwards (where the games do
a 3D engine with real perspective, and we fake perspective by making
backsprites 1.5x size).

This is, of course, wrong for gens 1-4, where the games fake perspective
by drawing pokemon in backsprites bigger, but making the image sizes the
same. This fixes it so the front sprites are the same size as the back
sprites, using the same fake perspective system.