Move teams to gen 6
We now officially require a generation mark at the beginning of all
formats. Formats whose ID do not start with "gen" will have "[Gen 6]"
prepended to their name.

This commit will prepend "gen6" to all existing formats in the
teambuilder, so that e.g. "OU" and "[Gen 6] OU" will be the same format
in the teambuilder.