Parse HTML time
This is supposed to allow for locale aware implementation of /roomevents.
Timezones are confusing, and can be misread causing some people coming
for an event in wrong time. This is client-side part of locale
aware /roomevents.

<time> element is part of HTML5 and is already parsed by Caja. Reusing
already existing element doesn't appear to be a practical issue, as it's
pretty much meant to do stuff like that.