Use different Pokemon in tests to avoid failures
I would prefer if the test would still test if a mega gets an item,
which is why Charizard-Mega-Y was picked. It's unlikely to be banned,
but is powerful enough to stay in OU.

A safer choice would be an Arceus, but there is no code yet that
automatically gives a correct item to formes (may be worth fixing).