Split 'activate' and 'upkeep' message types
Weather uses both 'activate':

    The mysterious strong winds weakened the attack!

and 'upkeep':

    The sandstorm is raging.

So keeping them on the same message type is complicated. PS already
uses `|upkeep|` to mark the beginning of the residual phase, so this is
a good name for this message type.

For now, only upkeep messages happening during the residual step are
here - Hail, Sandstorm, and Uproar. Messages that happen at other
times, such as Attract's "[POKEMON] is in love with [TARGET]!" remain
as "activate".