Fix trainercard alignment
Clicking on a user with a very large trainercard (generally, a bot in
a lot of rooms) would often push the trainercard offscreen even if it
fit, or push the trainercard to unscrollable areas.

With this change, the trainercard will no longer go offscreen unless
it doesn't fit, and if it doesn't fit, it will always be scrollable.